backend/Invoice/serializers.py
from rest_framework import serializers
from .models import Invoice

class InvoiceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Invoice
        fields = ('pk', 'customer', 'invoice_number', 'date', 'total_amount')

